refactor(PieceInfo): 담고 있는 정보에 맞는 패키지로 위치 이동
- PieceInfo 객체를 호출되는 클래스의 패키지 위치가 아닌, 실제 보유하고 있는 데이터가 표현하는 도메인 영역을 고려한
패키지 위치로 재배치함
